feat: improve `install` error when app doesnt exist
When trying to install an app that doesnt exist the error message
would advise the user that the app doesnt exist and that they should
run `install --help` to see what is available.

However, there may be situations where the user is using the wrong verb
because they are trying to get a tool, rather than an app.  This change introduces
a check when an app not found error is encountered to see whether a tool with the
user provided name exists and advises them accordingly.

also adds associated test.

Signed-off-by: Richard Gee <richard@technologee.co.uk>